+++ packages/desktop-client/src/components/budget/BalanceWithCarryover.tsx
@@ -1,6 +1,7 @@
 import React, { type ComponentProps } from 'react';
 
 import ArrowThinRight from '../../icons/v1/ArrowThinRight';
+import { type CSSProperties } from '../../style';
 import View from '../common/View';
 import CellValue from '../spreadsheet/CellValue';
 import useSheetValue from '../spreadsheet/useSheetValue';
@@ -11,17 +12,21 @@ type BalanceWithCarryoverProps = {
   carryover: ComponentProps<typeof CellValue>['binding'];
   balance: ComponentProps<typeof CellValue>['binding'];
   disabled?: boolean;
+  style?: CSSProperties;
+  carryoverStyle?: CSSProperties;
 };
 export default function BalanceWithCarryover({
   carryover,
   balance,
   disabled,
+  style,
+  carryoverStyle,
 }: BalanceWithCarryoverProps) {
   let carryoverValue = useSheetValue(carryover);
   let balanceValue = useSheetValue(balance);
 
   return (
-    <>
+    <View style={style}>
       <CellValue
         binding={balance}
         type="financial"
@@ -32,9 +37,10 @@ export default function BalanceWithCarryover({
             cursor: 'pointer',
             ':hover': { textDecoration: 'underline' },
           }),
+          ...style,
         }}
       />
-      {carryoverValue === true && (
+      {carryoverValue && (
         <View
           style={{
             alignSelf: 'center',
@@ -44,6 +50,7 @@ export default function BalanceWithCarryover({
             top: 0,
             bottom: 0,
             justifyContent: 'center',
+            ...carryoverStyle,
           }}
         >
           <ArrowThinRight
@@ -53,6 +60,6 @@ export default function BalanceWithCarryover({
           />
         </View>
       )}
-    </>
+    </View>
   );
 }
