--- packages/desktop-client/src/components/budget/CategoryGroupsContext.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export let CategoryGroupsContext = createContext([]);
