+++ packages/loot-core/src/types/models/category-group.d.ts
@@ -4,4 +4,6 @@ export interface CategoryGroupEntity {
   is_income?: boolean;
   sort_order?: number;
   tombstone?: boolean;
+  hidden?: boolean;
+  // categories?: CategoryEntity[];
 }
