+++ packages/loot-core/src/types/models/category.d.ts
@@ -1,10 +1,9 @@
-import type { CategoryGroupEntity } from './category-group';
-
 export interface CategoryEntity {
   id?: string;
   name: string;
   is_income?: boolean;
-  group: CategoryGroupEntity;
+  cat_group?: string;
   sort_order?: number;
   tombstone?: boolean;
+  hidden?: boolean;
 }
